LPS-55355 Move to proper folders
