+++ modules/apps/shopping/shopping-service/src/com/liferay/shopping/upgrade/v1_0_0/UpgradeShopping.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.shopping.upgrade;
+package com.liferay.shopping.upgrade.v1_0_0;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.shopping.upgrade.v1_0_0.ShoppingCartTable;
-import com.liferay.shopping.upgrade.v1_0_0.ShoppingOrderItemTable;
-import com.liferay.shopping.upgrade.v1_0_0.ShoppingOrderTable;
+import com.liferay.shopping.upgrade.v1_0_0.util.ShoppingCartTable;
+import com.liferay.shopping.upgrade.v1_0_0.util.ShoppingOrderItemTable;
+import com.liferay.shopping.upgrade.v1_0_0.util.ShoppingOrderTable;
 
 import java.sql.SQLException;
