+++ modules/apps/shopping/shopping-service/src/com/liferay/shopping/upgrade/v1_0_0/util/ShoppingOrderTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.shopping.upgrade.v1_0_0;
+package com.liferay.shopping.upgrade.v1_0_0.util;
 
 import java.sql.Types;
